modified output of poet (K=0 or not), update example of poet
Note:
1. cov(F) != I, due to the lack of "* (n-1) / n" staff

2. C range can be inspired by poet_Cmin, especially for "hard" thres, even
though it might not perform well (havn't try, seq(.5, 1.5, .1) is absolutely
wrong, see Figure 1 in An Overview on the Estimation of Large Covariance and
Precision Matrices)

3. output of poet and poet2 is changed to adapt the K = 0 and K > 0. Note that
for K = 0, we still output SigmaU, even though it is equal to SigmaY, this will
be helpful when we write poet_Cmin!

4. I am worried about "mat F, LamPCA;" in src/meth_poet.cpp, but I cannot find a
better way to store the F and LamPCA, which only occurs when K > 0.